Check Permission before action, Add Role to User

+++ app/src/main/java/com/msgsystems/jbugger/echipa2/backend/controllers/RoleController.java
@@ -1,12 +1,16 @@
 package com.msgsystems.jbugger.echipa2.backend.controllers;
 
+import com.msgsystems.jbugger.echipa2.backend.auth.AuthenticationService;
+import com.msgsystems.jbugger.echipa2.backend.auth.NotPermittedException;
 import com.msgsystems.jbugger.echipa2.backend.domain.Role;
 import com.msgsystems.jbugger.echipa2.backend.model.RolePermissionPairDTO;
 import com.msgsystems.jbugger.echipa2.backend.service.PermissionService;
 import com.msgsystems.jbugger.echipa2.backend.service.RoleService;
 import com.msgsystems.jbugger.echipa2.backend.service.ServiceOperationException;
+import com.msgsystems.jbugger.echipa2.backend.utils.constants.PermissionTypes;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,18 +21,19 @@ import java.util.List;
 public class RoleController {
     @Autowired
     RoleService roleService;
-
     @Autowired
     PermissionService permissionService;
+    @Autowired
+    AuthenticationService authService;
 
-    @RequestMapping(value = "/roles", method = RequestMethod.GET)
+    @GetMapping(value = "/roles")
     public ResponseEntity<List<Role>> findAllRoles(
             @RequestHeader(name="Authorization") String token
     ) throws ServiceOperationException {
       return roleService.findAllRoles().toResponseEntity();
     }
 
-    @RequestMapping(value = "/roles/permissions", method = RequestMethod.PUT)
+    @PutMapping(value = "/roles/permissions")
     public ResponseEntity<Role> addPermissionToRole(
             @RequestHeader(name="Authorization") String token,
             @RequestBody RolePermissionPairDTO body
@@ -40,12 +45,12 @@ public class RoleController {
                 .toResponseEntity();
     }
 
-    @RequestMapping(value = "/roles/permissions", method = RequestMethod.DELETE)
+    @DeleteMapping("/roles/permissions")
     public ResponseEntity<Role> removePermissionToRole(
-            @RequestHeader(name="Authorization") String token,
+            Authentication auth,
             @RequestBody RolePermissionPairDTO body
-    ) throws ServiceOperationException {
-        // TO DO: validate user privileges
+    ) throws ServiceOperationException, NotPermittedException {
+        authService.assertPermission(auth, PermissionTypes.PERMISSION_MANAGEMENT);
         var permission = permissionService.findByType(body.getPermissionType()).getValueInterceptError();
         var role = roleService.findByType(body.getRoleType()).getValueInterceptError();
         return roleService.removePermissionFromRole(role, permission)
